requirements.txt
fastapi
uvicorn
sqlalchemy
psycopg2-binary
python-dotenv
sqlmodel